+++ eslint.config.mjs
@@ -1,12 +1,18 @@
-import js from '@eslint/js'
+import eslint from '@eslint/js'
 import tseslint from 'typescript-eslint'
 
 export default tseslint.config(
-  js.configs.recommended,
-  ...tseslint.configs.recommended,
-  { ignores: ['*.*', 'scripts', 'out/**'] },
+  eslint.configs.recommended,
+  tseslint.configs.recommended,
+  {
+    ignores: ['*.*', 'out/**'],
+  },
   {
     files: ['src/**/*.ts'],
+    languageOptions: {
+      ecmaVersion: 2022,
+      sourceType: 'module',
+    },
     rules: {
       'no-empty': 'off',
       'no-extra-semi': 'off',